+++ android/java/org/chromium/chrome/browser/app/BraveActivity.java
@@ -217,17 +217,17 @@ public abstract class BraveActivity extends ChromeActivity
                    MiscAndroidMetricsConnectionErrorHandler
                            .MiscAndroidMetricsConnectionErrorHandlerDelegate {
     public static final String BRAVE_WALLET_HOST = "wallet";
-    public static final String BRAVE_WALLET_URL = "brave://wallet/crypto/portfolio/assets";
-    public static final String BRAVE_BUY_URL = "brave://wallet/crypto/fund-wallet";
-    public static final String BRAVE_SEND_URL = "brave://wallet/send";
-    public static final String BRAVE_SWAP_URL = "brave://wallet/swap";
-    public static final String BRAVE_DEPOSIT_URL = "brave://wallet/crypto/deposit-funds";
-    public static final String BRAVE_REWARDS_SETTINGS_URL = "brave://rewards/";
+    public static final String BRAVE_WALLET_URL = "aiwize://wallet/crypto/portfolio/assets";
+    public static final String BRAVE_BUY_URL = "aiwize://wallet/crypto/fund-wallet";
+    public static final String BRAVE_SEND_URL = "aiwize://wallet/send";
+    public static final String BRAVE_SWAP_URL = "aiwize://wallet/swap";
+    public static final String BRAVE_DEPOSIT_URL = "aiwize://wallet/crypto/deposit-funds";
+    public static final String BRAVE_REWARDS_SETTINGS_URL = "aiwize://rewards/";
     public static final String BRAVE_REWARDS_SETTINGS_WALLET_VERIFICATION_URL =
-            "brave://rewards/#verify";
-    public static final String BRAVE_REWARDS_WALLET_RECONNECT_URL = "brave://rewards/reconnect";
-    public static final String BRAVE_REWARDS_SETTINGS_MONTHLY_URL = "brave://rewards/#monthly";
-    public static final String REWARDS_AC_SETTINGS_URL = "brave://rewards/contribute";
+            "aiwize://rewards/#verify";
+    public static final String BRAVE_REWARDS_WALLET_RECONNECT_URL = "aiwize://rewards/reconnect";
+    public static final String BRAVE_REWARDS_SETTINGS_MONTHLY_URL = "aiwize://rewards/#monthly";
+    public static final String REWARDS_AC_SETTINGS_URL = "aiwize://rewards/contribute";
     public static final String BRAVE_AI_CHAT_URL = "chrome-untrusted://chat";
     public static final String REWARDS_LEARN_MORE_URL =
             "https://brave.com/faq-rewards/#unclaimed-funds";