+++ browser/brave_wallet/solana_provider_renderer_browsertest.cc
@@ -569,7 +569,7 @@ class SolanaProviderRendererTest : public InProcessBrowserTest {
     // This is intentional to trigger
     // TestBraveContentBrowserClient::RegisterBrowserInterfaceBindersForFrame
     ASSERT_TRUE(
-        ui_test_utils::NavigateToURL(browser(), GURL("brave://settings")));
+        ui_test_utils::NavigateToURL(browser(), GURL("aiwize://settings")));
 
     GURL url = embedded_test_server()->GetURL("/empty.html");
     ASSERT_TRUE(ui_test_utils::NavigateToURL(browser(), url));