+++ browser/tor/onion_location_navigation_throttle_browsertest.cc
@@ -40,7 +40,7 @@ namespace {
 constexpr char kTestOnionPath[] = "/onion";
 constexpr char kTestOnionURL[] = "https://brave.onion";
 constexpr char kTestInvalidScheme[] = "/invalid_scheme";
-constexpr char kTestInvalidSchemeURL[] = "brave://brave.onion";
+constexpr char kTestInvalidSchemeURL[] = "aiwize://brave.onion";
 constexpr char kTestNotOnion[] = "/not_onion";
 constexpr char kTestNotOnionURL[] = "https://brave.com";
 constexpr char kTestErrorPage[] = "/errorpage";