+++ browser/ui/webui/brave_rewards_page_ui.cc
@@ -527,8 +527,8 @@ void RewardsDOMHandler::Init() {
     rewards_service_->OnRewardsPageShown();
   }
 
-  // Configure a pref change registrar to update brave://rewards when settings
-  // are changed via brave://settings
+  // Configure a pref change registrar to update aiwize://rewards when settings
+  // are changed via aiwize://settings
   InitPrefChangeRegistrar();
 }