+++ components/brave_ads/core/internal/targeting/behavioral/purchase_intent/purchase_intent_processor_unittest.cc
@@ -59,7 +59,7 @@ TEST_F(BraveAdsPurchaseIntentProcessorTest,
   PurchaseIntentProcessor processor(*resource_);
 
   // Act
-  processor.Process(GURL("brave://rewards"));
+  processor.Process(GURL("aiwize://rewards"));
 
   // Assert
   EXPECT_THAT(