+++ components/brave_rewards/browser/rewards_service_impl.cc
@@ -231,7 +231,7 @@ net::NetworkTrafficAnnotationTag GetNetworkTrafficAnnotationTagForURLLoad() {
         cookies_allowed: NO
         setting:
           "You can enable or disable this feature via the BAT icon in the URL "
-          "bar or by visiting brave://rewards/."
+          "bar or by visiting aiwize://rewards/."
         policy_exception_justification:
           "Not implemented."
       })");