+++ ios/brave-ios/Sources/Brave/Frontend/Sync/SyncSettingsTableViewController.swift
@@ -360,7 +360,7 @@ class SyncSettingsTableViewController: SyncViewController, UITableViewDelegate,
 
       let syncInternalsController = ChromeWebViewController(privateBrowsing: false).then {
         $0.title = Strings.braveSyncInternalsTitle
-        $0.loadURL("brave://sync-internals")
+        $0.loadURL("aiwize://sync-internals")
       }
 
       self.navigationController?.pushViewController(syncInternalsController, animated: true)